app/requirements.txt
Flask==2.1.1
mysql-connector-python==8.0.28
PyYAML==5.3.1
requests==2.26.0
tinydb==4.8.0
Werkzeug==2.2.2
influxdb-client==1.45.0
docker==7.1.0